YilaoServer/requirements.txt
Flask~=1.1.2
mkdocs~=1.1.2
marshmallow~=3.7.1
mysql-connector-python~=8.0.21
redis~=3.5.3
requests~=2.24.0
Werkzeug~=1.0.1
aliyun-python-sdk-core~=2.13.30
flask_restful~=0.3.8
gunicorn~=20.0.4
Shapely~=1.7.1
